Migrated to mysql with all tests green

+++ src/Inshapardaz.Domain/Adapters/Repositories/IFileRepository.cs
@@ -6,12 +6,12 @@ namespace Inshapardaz.Domain.Repositories
 {
     public interface IFileRepository
     {
-        Task<FileModel> GetFileById(int id, CancellationToken cancellationToken);
+        Task<FileModel> GetFileById(long id, CancellationToken cancellationToken);
 
         Task<FileModel> AddFile(FileModel file, CancellationToken cancellationToken);
 
         Task UpdateFile(FileModel file, CancellationToken cancellationToken);
 
-        Task DeleteFile(int id, CancellationToken cancellationToken);
+        Task DeleteFile(long id, CancellationToken cancellationToken);
     }
 }
