+++ src/Inshapardaz.Domain/Adapters/Repositories/Library/IBookPageRepository.cs
@@ -12,13 +12,13 @@ namespace Inshapardaz.Domain.Adapters.Repositories.Library
 
         Task<int> GetLastPageNumberForBook(int libraryId, int bookId, CancellationToken cancellationToken);
 
-        Task<BookPageModel> AddPage(int libraryId, int bookId, int sequenceNumber, string text, int imageId, int? chapterId, CancellationToken cancellationToken);
+        Task<BookPageModel> AddPage(int libraryId, int bookId, int sequenceNumber, string text, long imageId, long? chapterId, CancellationToken cancellationToken);
 
-        Task<BookPageModel> UpdatePage(int libraryId, int bookId, int sequenceNumber, string text, int imageId, EditingStatus status, int? chapterId, CancellationToken cancellationToken);
+        Task<BookPageModel> UpdatePage(int libraryId, int bookId, int sequenceNumber, string text, long imageId, EditingStatus status, long? chapterId, CancellationToken cancellationToken);
 
         Task DeletePage(int libraryId, int bookId, int sequenceNumber, CancellationToken cancellationToken);
 
-        Task<BookPageModel> UpdatePageImage(int libraryId, int bookId, int sequenceNumber, int imageId, CancellationToken cancellationToken);
+        Task<BookPageModel> UpdatePageImage(int libraryId, int bookId, int sequenceNumber, long imageId, CancellationToken cancellationToken);
 
         Task<IEnumerable<BookPageModel>> GetAllPagesByBook(int libraryId, int bookId, CancellationToken cancellationToken);
 
@@ -27,7 +27,7 @@ namespace Inshapardaz.Domain.Adapters.Repositories.Library
         Task<int> GetPageCount(int libraryId, int bookId, int oldSequenceNumber, CancellationToken cancellationToken);
         Task<Page<BookPageModel>> GetPagesByBook(int libraryId, int bookId, int pageNumber, int pageSize, EditingStatus status, AssignmentFilter assignmentFilter, AssignmentFilter reviewerAssignmentFilter, int? assignedTo, CancellationToken cancellationToken);
         Task<Page<BookPageModel>> GetPagesByUser(int libraryId, int assignedTo, EditingStatus statusFilter, int pageNumber, int pageSize, CancellationToken cancellationToken);
-        Task<IEnumerable<BookPageModel>> GetPagesByBookChapter(int libraryId, int bookId, int chapterId, CancellationToken cancellationToken);
+        Task<IEnumerable<BookPageModel>> GetPagesByBookChapter(int libraryId, int bookId, long chapterId, CancellationToken cancellationToken);
 
         Task<BookPageModel> UpdateWriterAssignment(int libraryId, int bookId, int sequenceNumber, int? assignedAccountId, CancellationToken cancellationToken);
         Task UpdatePageSequenceNumber(int libraryId, int bookId, int oldSequenceNumber, int newSequenceNumber, CancellationToken cancellationToken);
