+++ src/Inshapardaz.Domain/Adapters/Repositories/Library/ILibraryRepository.cs
@@ -38,6 +38,6 @@ namespace Inshapardaz.Domain.Adapters.Repositories.Library
         Task RemoveLibraryFromAccount(int libraryId, int accountId, CancellationToken cancellationToken);
 
         Task<IEnumerable<LibraryModel>> GetLibrariesByAccountId(int accountId, CancellationToken cancellationToken = default(CancellationToken));
-        Task UpdateLibraryImage(int libraryId, int imageId, CancellationToken cancellationToken);
+        Task UpdateLibraryImage(int libraryId, long imageId, CancellationToken cancellationToken);
     }
 }
