+++ src/Inshapardaz.Domain/Adapters/Repositories/Library/IPeriodicalRepository.cs
@@ -16,6 +16,6 @@ namespace Inshapardaz.Domain.Repositories.Library
         Task<PeriodicalModel> UpdatePeriodical(int libraryId, PeriodicalModel periodical, CancellationToken cancellationToken);
 
         Task DeletePeriodical(int libraryId, int periodicalId, CancellationToken cancellationToken);
-        Task UpdatePeriodicalImage(int libraryId, int periodicalId, int imageId, CancellationToken cancellationToken);
+        Task UpdatePeriodicalImage(int libraryId, int periodicalId, long imageId, CancellationToken cancellationToken);
     }
 }
