+++ src/Inshapardaz.Domain/Models/Library/BookPageModel.cs
@@ -7,7 +7,7 @@ namespace Inshapardaz.Domain.Models.Library
         public string Text { get; set; }
         public int SequenceNumber { get; set; }
         public int BookId { get; set; }
-        public int? ImageId { get; set; }
+        public long? ImageId { get; set; }
         public string ImageUrl { get; set; }
         public EditingStatus Status { get; set; }
         public int? WriterAccountId { get; set; }
@@ -16,7 +16,7 @@ namespace Inshapardaz.Domain.Models.Library
         public int? ReviewerAccountId { get; set; }
         public string ReviewerAccountName { get; set; }
         public DateTime? ReviewerAssignTimeStamp { get; set; }
-        public int? ChapterId { get; set; }
+        public long? ChapterId { get; set; }
 
         public string ChapterTitle { get; set; }
         public BookPageModel PreviousPage { get; internal set; }
