+++ src/Inshapardaz.Domain/Models/Library/IssueModel.cs
@@ -13,7 +13,7 @@ namespace Inshapardaz.Domain.Models.Library
 
         public DateTime IssueDate { get; set; }
 
-        public int? ImageId { get; set; }
+        public long? ImageId { get; set; }
 
         public int PeriodicalId { get; set; }
