+++ src/Inshapardaz.Domain/Models/Library/PeriodicalModel.cs
@@ -11,7 +11,7 @@ namespace Inshapardaz.Domain.Models.Library
 
         public string Description { get; set; }
 
-        public int? ImageId { get; set; }
+        public long? ImageId { get; set; }
 
         public string ImageUrl { get; set; }
